+++ Libplanet.Explorer/GraphTypes/TxResultType.cs
@@ -7,8 +7,6 @@ namespace Libplanet.Explorer.GraphTypes
     {
         public TxResultType()
         {
-            Name = "TxResult";
-
             Field<NonNullGraphType<TxStatusType>>(
                 nameof(TxResult.TxStatus),
                 description: "The transaction status.",
@@ -27,14 +25,14 @@ namespace Libplanet.Explorer.GraphTypes
                 resolve: context => context.Source.BlockHash
             );
 
-            Field<HashDigestType<SHA256>>(
+            Field<HashDigestSHA256Type>(
                 nameof(TxResult.InputState),
                 description: "The input state's root hash " +
                 "which the target transaction executed.",
                 resolve: context => context.Source.InputState
             );
 
-            Field<HashDigestType<SHA256>>(
+            Field<HashDigestSHA256Type>(
                 nameof(TxResult.OutputState),
                 description: "The output state's root hash " +
                 "which the target transaction executed.",